refactor: move auth components to pages folder for auth
