+++ apps/web/src/pages/auth/PasswordResetPage.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { Center } from '@mantine/core';
 import AuthLayout from '../../components/layout/components/AuthLayout';
 import AuthContainer from '../../components/layout/components/AuthContainer';
-import { PasswordResetRequestForm } from '../../components/auth/PasswordResetRequestForm';
-import { PasswordResetForm } from '../../components/auth/PasswordResetForm';
+import { PasswordResetRequestForm } from './components/PasswordResetRequestForm';
+import { PasswordResetForm } from './components/PasswordResetForm';
 import { Button, Text } from '../../design-system';
 import { ROUTES } from '../../constants/routes.enum';
 import { useVercelParams } from '../../hooks';
