+++ apps/web/src/pages/auth/SignUpPage.tsx
@@ -1,5 +1,5 @@
 import AuthLayout from '../../components/layout/components/AuthLayout';
-import { SignUpForm } from '../../components/auth/SignUpForm';
+import { SignUpForm } from './components/SignUpForm';
 import AuthContainer from '../../components/layout/components/AuthContainer';
 import { useBlueprint } from '../../hooks';
