+++ apps/web/src/pages/auth/components/CreateOrganizationForm.tsx
@@ -5,11 +5,11 @@ import { useNavigate } from 'react-router-dom';
 import decode from 'jwt-decode';
 import { IJwtPayload } from '@novu/shared';
 
-import { Button, Input } from '../../design-system';
-import { api } from '../../api/api.client';
-import { useAuthContext } from '../providers/AuthProvider';
-import { useVercelIntegration, useVercelParams } from '../../hooks';
-import { ROUTES } from '../../constants/routes.enum';
+import { Button, Input } from '../../../design-system';
+import { api } from '../../../api/api.client';
+import { useAuthContext } from '../../../components/providers/AuthProvider';
+import { useVercelIntegration, useVercelParams } from '../../../hooks';
+import { ROUTES } from '../../../constants/routes.enum';
 
 type Props = {};
