+++ apps/web/src/pages/auth/components/LoginForm.tsx
@@ -6,15 +6,15 @@ import { useForm } from 'react-hook-form';
 import * as Sentry from '@sentry/react';
 import { Divider, Button as MantineButton, Center } from '@mantine/core';
 
-import { useAuthContext } from '../providers/AuthProvider';
-import { api } from '../../api/api.client';
-import { PasswordInput, Button, colors, Input, Text } from '../../design-system';
-import { GitHub } from '../../design-system/icons';
-import { IS_DOCKER_HOSTED } from '../../config';
-import { useVercelParams } from '../../hooks';
+import { useAuthContext } from '../../../components/providers/AuthProvider';
+import { api } from '../../../api/api.client';
+import { PasswordInput, Button, colors, Input, Text } from '../../../design-system';
+import { GitHub } from '../../../design-system/icons';
+import { IS_DOCKER_HOSTED } from '../../../config';
+import { useVercelParams } from '../../../hooks';
 import { useAcceptInvite } from './useAcceptInvite';
 import { buildGithubLink, buildVercelGithubLink } from './gitHubUtils';
-import { ROUTES } from '../../constants/routes.enum';
+import { ROUTES } from '../../../constants/routes.enum';
 
 type LoginFormProps = {
   invitationToken?: string;
