+++ apps/web/src/pages/auth/components/PasswordRequirementPopover.tsx
@@ -1,9 +1,9 @@
 import { useState, useMemo } from 'react';
 import { Text, Box, createStyles, Popover } from '@mantine/core';
 import { useWatch } from 'react-hook-form';
-import { Check } from '../../design-system/icons';
-import { Close } from '../../design-system/icons/actions/Close';
-import { colors } from '../../design-system';
+import { Check } from '../../../design-system/icons';
+import { Close } from '../../../design-system/icons/actions/Close';
+import { colors } from '../../../design-system';
 import { PasswordStrengthBar } from './PasswordStrengthBar';
 
 const usePopoverStyles = createStyles(({ colorScheme }) => ({
