+++ apps/web/src/pages/auth/components/PasswordResetRequestForm.tsx
@@ -4,10 +4,10 @@ import { Link } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { Center } from '@mantine/core';
 
-import { api } from '../../api/api.client';
-import { Button, colors, Input, Text } from '../../design-system';
-import { useVercelParams } from '../../hooks';
-import { ROUTES } from '../../constants/routes.enum';
+import { api } from '../../../api/api.client';
+import { Button, colors, Input, Text } from '../../../design-system';
+import { useVercelParams } from '../../../hooks';
+import { ROUTES } from '../../../constants/routes.enum';
 
 type Props = {
   onSent: () => void;
