+++ apps/web/src/pages/auth/components/PasswordStrengthBar.tsx
