+++ apps/web/src/pages/auth/components/SetupLoader.tsx
@@ -1,5 +1,5 @@
 import { Loader, Paper } from '@mantine/core';
-import { colors, Text } from '../../design-system';
+import { colors, Text } from '../../../design-system';
 
 type Props = {
   title: string;
